requirements.txt
# Core dependencies
psutil>=5.9.0
wmi>=1.5.1
pywin32>=305
requests>=2.28.0
colorama>=0.4.6
pandas>=2.0.0

# Optional network monitoring
scapy>=2.5.0
netifaces>=0.11.0

# Environment configuration
python-dotenv>=1.0.0
